backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knotwork"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    # Validation
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # LangGraph (core only — no LangChain LLM wrappers after S7)
    "langgraph>=0.2",
    "langgraph-checkpoint-postgres>=2.0",
    "langchain-core>=0.3",
    # First-party LLM SDKs (adapters use these directly)
    "anthropic>=0.40",
    "openai>=1.50",
    # Queue
    "arq>=0.26",
    "redis>=5.0",
    # Storage
    "boto3>=1.34",
    "aiofiles>=23.2",
    # Token counting
    "tiktoken>=0.7",
    # HTTP client
    "httpx>=0.27",
    # Notifications
    "python-telegram-bot>=21.0",
    "sendgrid>=6.11",
    # MCP
    "mcp>=1.0",
    # Handbook file conversion (S6.5)
    "pypdf>=4.0",
    "python-docx>=1.1",
    "markdownify>=0.11",
    # .doc (Word 97-2003) extraction
    "olefile>=0.47",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "factory-boy>=3.3",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
